+++ week02/drone_26_9.py
