+++ week03/drone_27_9.py
