+++ week06/drone_13_10.py
