+++ week06/telemetry_18_10.py
