+++ week09/friday/drone.py
